task №2/requirements.txt
playwright>=1.40.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
